Use commonly known project structures
It is best practice to divide a code repository in the following known sub folders (where applicable):
* src    -- short for "source". Here's the location of the source code.
* doc   -- short for "documentation". Here's the location of the documentation.
* bin    -- short for "binaries". Here's the location for executable program files.
